--- backend.native/compiler/ir/backend.native/src/org/jetbrains/kotlin/backend/konan/library/KonanLibraryReader.kt
@@ -1,43 +1,0 @@
-/*
- * Copyright 2010-2017 JetBrains s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jetbrains.kotlin.backend.konan.library
-
-import org.jetbrains.kotlin.config.LanguageVersionSettings
-import org.jetbrains.kotlin.descriptors.ModuleDescriptor
-import org.jetbrains.kotlin.konan.properties.Properties
-
-interface KonanLibraryReader {
-    val libraryName: String
-    val uniqueName: String
-    val bitcodePaths: List<String>
-    val includedPaths: List<String>
-    val linkerOpts: List<String>
-    val unresolvedDependencies: List<String>
-    val dataFlowGraph: ByteArray?
-    val isDefaultLibrary: Boolean get() = false
-    val isNeededForLink: Boolean get() = true
-    val manifestProperties: Properties
-    val moduleHeaderData: ByteArray
-    fun packageMetadata(fqName: String): ByteArray
-    fun markPackageAccessed(fqName: String)
-    fun moduleDescriptor(specifics: LanguageVersionSettings): ModuleDescriptor
-}
-
-interface MetadataReader {
-    fun loadSerializedModule(): ByteArray
-    fun loadSerializedPackageFragment(fqName: String): ByteArray
-}
