--- extracted/konan.metadata.serializer/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = "konan.metadata.serializer"
