+++ extracted/konan.serializer/build.gradle
@@ -19,7 +19,7 @@ repositories {
 
 sourceSets {
     main.kotlin {
-        srcDir 'src/main/kotlin'
+        srcDir 'src'
     }
 }
