+++ extracted/konan.serializer/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = "konan.serializer"
