+++ extracted/konan.serializer/src/org/jetbrains/kotlin/konan/library/KonanMetadataReader.kt
@@ -1,0 +1,6 @@
+package org.jetbrains.kotlin.konan.library
+
+interface MetadataReader {
+    fun loadSerializedModule(): ByteArray
+    fun loadSerializedPackageFragment(fqName: String): ByteArray
+}
