+++ extracted/konan.serializer/src/org/jetbrains/kotlin/serialization/konan/KonanSerializerProtocol.kt
@@ -1,0 +1,21 @@
+package org.jetbrains.kotlin.serialization.konan
+
+import org.jetbrains.kotlin.metadata.konan.KonanProtoBuf
+import org.jetbrains.kotlin.protobuf.ExtensionRegistryLite
+import org.jetbrains.kotlin.serialization.SerializerExtensionProtocol
+
+object KonanSerializerProtocol : SerializerExtensionProtocol(
+        ExtensionRegistryLite.newInstance().apply {
+            KonanProtoBuf.registerAllExtensions(this)
+        },
+        KonanProtoBuf.packageFqName,
+        KonanProtoBuf.constructorAnnotation,
+        KonanProtoBuf.classAnnotation,
+        KonanProtoBuf.functionAnnotation,
+        KonanProtoBuf.propertyAnnotation,
+        KonanProtoBuf.enumEntryAnnotation,
+        KonanProtoBuf.compileTimeValue,
+        KonanProtoBuf.parameterAnnotation,
+        KonanProtoBuf.typeAnnotation,
+        KonanProtoBuf.typeParameterAnnotation
+)
